Updated all types and split them into their own files
