+++ src/components/Card.svelte
@@ -1,37 +1,35 @@
 <script lang="ts">
-	import { fade, fly, slide } from 'svelte/transition';
+	import { fly } from 'svelte/transition';
 	export let title: string;
 	export let width = 'w-full';
 	export let disableToggle = false;
 	export let showOnLoad = true;
-	let showTerms = showOnLoad;
-	const toggleTerms = () => {
-		showTerms = !showTerms;
+	let showContent = showOnLoad;
+	const toggleContent = () => {
+		showContent = !showContent;
 	};
 </script>
 
-<section in:fly={{ x: -200 }} class="card shadow-md h-fit  {width} p-4">
-	<div class="flex flow-row justify-between my-1">
-		<h4 class="text-md">{title}</h4>
+<section in:fly={{ x: -200 }} class="card shadow-md h-fit  {width} p-4 gap-4">
+	<div class="flex flow-row justify-between">
+		<p class="text-md">{title}</p>
 
 		{#if !disableToggle}
-			<div>
-				<button on:click={toggleTerms} class="toggle-btn text-center hover:underline">
-					{#if showTerms}
-						<i class="fa-solid fa-minus" />
-						<span>Hide</span>
-					{:else}
-						<i class="fa-solid fa-plus" />
-						<span>Show</span>
-					{/if}
-				</button>
-			</div>
+			<button on:click={toggleContent} class="toggle-btn text-center hover:underline">
+				{#if showContent}
+					<i class="fa-solid fa-minus" />
+					<span>Hide</span>
+				{:else}
+					<i class="fa-solid fa-plus" />
+					<span>Show</span>
+				{/if}
+			</button>
 		{/if}
 	</div>
 
-	{#if showTerms || disableToggle}
-		<div>
+	<div>
+		{#if showContent || disableToggle}
 			<slot />
-		</div>
-	{/if}
+		{/if}
+	</div>
 </section>
