+++ src/routes/arc/+page.svelte
@@ -3,9 +3,5 @@
 </script>
 
 <main>
-	<div class="grid grid-cols-1 md:grid-cols-1  lg:grid-cols-12 ">
-		<div class={`col-span-1 md:col-span-1 lg:col-span-12`}>
-			<HighResMap />
-		</div>
-	</div>
+	<HighResMap />
 </main>
